requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6
langchain>=0.0.350
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-huggingface>=0.0.10
chromadb>=0.4.22
transformers>=4.35.2
torch>=2.1.1
sentencepiece>=0.1.99
sentence-transformers>=2.2.2
pyyaml>=6.0.1
python-dotenv>=1.0.0
tavily-python>=0.3.0
langgraph>=0.0.15
langchain_ollama>=0.0.10